More work on serializing, implemented all elements/objects.

+++ api/src/main/java/oxy/bascenario/api/render/elements/emoticon/Emoticon.java
@@ -3,5 +3,5 @@ package oxy.bascenario.api.render.elements.emoticon;
 import lombok.Builder;
 
 @Builder(toBuilder = true, builderClassName = "Builder")
-public record Emoticon(long duration, EmoticonType type, boolean sound) {
+public record Emoticon(int duration, EmoticonType type, boolean sound) {
 }
